extract PredicatesAssert
